refactor: change structure of class and update methods
